backend/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
pymongo==4.16.0
motor==3.7.1
email-validator==2.3.0
python-dotenv==1.0.0
